+++ io_mesh_w3d/bfme/tools/destroy_animation.py
@@ -362,6 +362,7 @@ class DESTROY_PT_panel(Panel):
     bl_space_type = 'VIEW_3D'
     bl_region_type = 'UI'
     bl_category = 'W3D Tools'
+    bl_options = {'DEFAULT_CLOSED'}
 
     def draw(self, context):
         layout = self.layout