+++ io_mesh_w3d/bfme/tools/existing_animations.py
@@ -177,6 +177,7 @@ class EXISTING_ANIMATIONS_PT_panel(Panel):
     bl_space_type = 'VIEW_3D'
     bl_region_type = 'UI'
     bl_category = 'W3D Tools'
+    bl_options = {'DEFAULT_CLOSED'}
 
     def draw(self, context):
         layout = self.layout