+++ io_mesh_w3d/bfme/tools/export_settings.py
@@ -209,6 +209,7 @@ class BFME_PT_export_settings(Panel):
     bl_space_type = 'VIEW_3D'
     bl_region_type = 'UI'
     bl_category = 'W3D Tools'
+    bl_options = {'DEFAULT_CLOSED'}
 
     def draw(self, context):
         layout = self.layout